Move events-handler to src/utils
